# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Management Tool API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  member: "member",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  avatarUrl?: string | null;
  role: UserRole;
  createdAt: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  active: "active",
  archived: "archived",
  completed: "completed",
} as const;

export interface Project {
  id: number;
  name: string;
  description?: string | null;
  color: string;
  status: ProjectStatus;
  ownerId: number;
  startDate?: string | null;
  endDate?: string | null;
  taskCount: number;
  completedTaskCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectRequest {
  name: string;
  description?: string;
  color: string;
  startDate?: string;
  endDate?: string;
}

export type UpdateProjectRequestStatus =
  (typeof UpdateProjectRequestStatus)[keyof typeof UpdateProjectRequestStatus];

export const UpdateProjectRequestStatus = {
  active: "active",
  archived: "archived",
  completed: "completed",
} as const;

export interface UpdateProjectRequest {
  name?: string;
  description?: string;
  color?: string;
  status?: UpdateProjectRequestStatus;
  startDate?: string;
  endDate?: string;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  todo: "todo",
  in_progress: "in_progress",
  in_review: "in_review",
  done: "done",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  projectId: number;
  parentTaskId?: number | null;
  assigneeId?: number | null;
  reporterId: number;
  assignee?: User | null;
  reporter?: User | null;
  subtaskCount: number;
  blockedByIds: number[];
  blockingIds: number[];
  startDate?: string | null;
  dueDate?: string | null;
  estimatedHours?: number | null;
  tags: string[];
  position: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateTaskRequestStatus =
  (typeof CreateTaskRequestStatus)[keyof typeof CreateTaskRequestStatus];

export const CreateTaskRequestStatus = {
  todo: "todo",
  in_progress: "in_progress",
  in_review: "in_review",
  done: "done",
} as const;

export type CreateTaskRequestPriority =
  (typeof CreateTaskRequestPriority)[keyof typeof CreateTaskRequestPriority];

export const CreateTaskRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface CreateTaskRequest {
  title: string;
  description?: string;
  status?: CreateTaskRequestStatus;
  priority?: CreateTaskRequestPriority;
  assigneeId?: number;
  startDate?: string;
  dueDate?: string;
  estimatedHours?: number;
  tags?: string[];
}

export type CreateSubtaskRequestStatus =
  (typeof CreateSubtaskRequestStatus)[keyof typeof CreateSubtaskRequestStatus];

export const CreateSubtaskRequestStatus = {
  todo: "todo",
  in_progress: "in_progress",
  in_review: "in_review",
  done: "done",
} as const;

export type CreateSubtaskRequestPriority =
  (typeof CreateSubtaskRequestPriority)[keyof typeof CreateSubtaskRequestPriority];

export const CreateSubtaskRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface CreateSubtaskRequest {
  title: string;
  status?: CreateSubtaskRequestStatus;
  priority?: CreateSubtaskRequestPriority;
  assigneeId?: number;
  dueDate?: string;
}

export type UpdateTaskRequestStatus =
  (typeof UpdateTaskRequestStatus)[keyof typeof UpdateTaskRequestStatus];

export const UpdateTaskRequestStatus = {
  todo: "todo",
  in_progress: "in_progress",
  in_review: "in_review",
  done: "done",
} as const;

export type UpdateTaskRequestPriority =
  (typeof UpdateTaskRequestPriority)[keyof typeof UpdateTaskRequestPriority];

export const UpdateTaskRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface UpdateTaskRequest {
  title?: string;
  description?: string;
  status?: UpdateTaskRequestStatus;
  priority?: UpdateTaskRequestPriority;
  assigneeId?: number;
  startDate?: string;
  dueDate?: string;
  estimatedHours?: number;
  tags?: string[];
  position?: number;
}

export type TaskDependencyDependsOnTask = {
  id?: number;
  title?: string;
  status?: string;
} | null;

export interface TaskDependency {
  id: number;
  taskId: number;
  dependsOnTaskId: number;
  dependsOnTask?: TaskDependencyDependsOnTask;
  createdAt: string;
}

export interface AddDependencyRequest {
  dependsOnTaskId: number;
}

export interface Comment {
  id: number;
  content: string;
  taskId: number;
  userId: number;
  user?: User | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCommentRequest {
  content: string;
}

export type ListTasksParams = {
  status?: ListTasksStatus;
  assigneeId?: number;
  priority?: ListTasksPriority;
};

export type ListTasksStatus =
  (typeof ListTasksStatus)[keyof typeof ListTasksStatus];

export const ListTasksStatus = {
  todo: "todo",
  in_progress: "in_progress",
  in_review: "in_review",
  done: "done",
} as const;

export type ListTasksPriority =
  (typeof ListTasksPriority)[keyof typeof ListTasksPriority];

export const ListTasksPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;